vrrp: extract receiver/packet-parse to instance_receive.go
Pure code-motion split of pkg/vrrp/instance.go (#5661). Move the
inbound receive loops, AF_PACKET/IPv6 parsing, ifindex-arrival gating,
enqueue, and RX-drop warning into a cohesive sibling file, package
vrrp:

  isTimeoutError, acceptArrivalIfindex, expectedIfindex, receiver,
  receiverIPv6, receiverAfPacket, parseAfPacketIPv4, parseAfPacketIPv6,
  enqueuePacket, walkIPv6ExtHeaders, warnRXDrop.

Declarations moved verbatim — no rename of exported symbols, no
logic/timing/const change, and no reorder of any goroutine, defer,
channel, or mutex operation (whole receiver goroutines and their
select/lock/atomic sequences moved intact). Imports pruned by
goimports; gofmt clean. Builds green.
